refactoring load_data and open_list_archives from mailman.py to archive.py, fixes #435
